pyproject.toml
[project]
name = "aesop"
version = "0.1.0"
description = "Developer-first CLI for threat modeling AI agents, LLM applications, and RAG systems using MITRE ATLAS-inspired mappings and deterministic security analysis."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Muhammed Can" }]
keywords = ["security", "threat-modeling", "ai-agents", "llm", "mitre-atlas", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Security",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.12,<1",
    "pydantic>=2.6,<3",
    "pyyaml>=6.0,<7",
    "rich>=13.7,<14",
    "jinja2>=3.1,<4",
]

[project.urls]
Homepage = "https://github.com/muhammedcan/aesop"
Repository = "https://github.com/muhammedcan/aesop"
Issues = "https://github.com/muhammedcan/aesop/issues"

[project.scripts]
aesop = "aesop.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=8.0,<9"]
